+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please mark backwards incompatible changes with an exclamation mark at the start
 ## [Unreleased]
 
 ### Added
+- The `sort` parameter for Elasticsearch's `TopHits` aggregation.
 - The `missing` parameter for Elasticsearch's `Terms` aggregation.
 
 ## [29.10.0] - 2026-08-21